customize serial No or scan barcode
