--- goldapp/golds/doctype/gold_ledger/gold_ledger.py
@@ -1,8 +1,0 @@
-# Copyright (c) 2024, hevin and contributors
-# For license information, please see license.txt
-
-# import frappe
-from frappe.model.document import Document
-
-class GoldLedger(Document):
-	pass
