+++ goldapp/golds/doctype/gold_ledger_entry/gold_ledger_entry.js
@@ -1,7 +1,7 @@
 // Copyright (c) 2024, hevin and contributors
 // For license information, please see license.txt
 
-frappe.ui.form.on('Gold Ledger', {
+frappe.ui.form.on('Gold Ledger Entry', {
 	// refresh: function(frm) {
 
 	// }
