+++ goldapp/golds/doctype/gold_ledger_entry/test_gold_ledger_entry.py
@@ -5,5 +5,5 @@
 from frappe.tests.utils import FrappeTestCase
 
 
-class TestGoldLedger(FrappeTestCase):
+class TestGoldLedgerEntry(FrappeTestCase):
 	pass
