+++ goldapp/golds/doctype/metal_rate/metal_rate.json
@@ -29,7 +29,8 @@
    "fieldname": "daily_metal_rate",
    "fieldtype": "Table",
    "label": "Daily Metal Rate",
-   "options": "Daily Metal Rate"
+   "options": "Daily Metal Rate",
+   "reqd": 1
   },
   {
    "fieldname": "amended_from",
@@ -45,7 +46,7 @@
  "index_web_pages_for_search": 1,
  "is_submittable": 1,
  "links": [],
- "modified": "2024-03-18 11:42:38.302945",
+ "modified": "2024-04-03 11:35:40.344574",
  "modified_by": "Administrator",
  "module": "golds",
  "name": "Metal Rate",
