+++ goldapp/golds/report/gold/gold.json
@@ -1,0 +1,36 @@
+{
+ "add_total_row": 0,
+ "columns": [],
+ "creation": "2024-04-03 12:13:25.386302",
+ "disable_prepared_report": 0,
+ "disabled": 0,
+ "docstatus": 0,
+ "doctype": "Report",
+ "filters": [],
+ "idx": 0,
+ "is_standard": "Yes",
+ "modified": "2024-04-03 12:13:25.386302",
+ "modified_by": "Administrator",
+ "module": "golds",
+ "name": "gold",
+ "owner": "Administrator",
+ "prepared_report": 0,
+ "query": "select * from `tabStock Entry`;",
+ "ref_doctype": "Stock Entry",
+ "report_name": "gold",
+ "report_type": "Query Report",
+ "roles": [
+  {
+   "role": "Stock User"
+  },
+  {
+   "role": "Manufacturing User"
+  },
+  {
+   "role": "Manufacturing Manager"
+  },
+  {
+   "role": "Stock Manager"
+  }
+ ]
+}
