+++ goldapp/public/js/sales_order_item.js
@@ -1,21 +1,60 @@
-frappe.ui.form.on('Sales Order Item', {
-    item_code: fetchGoldRate,
-    custom_gold_type: fetchGoldRate,
-    custom_purity: fetchGoldRate, 
-    custom_gross_weight: updateItem,
-    custom_less_weight: updateItem,
-    custom_purity_percentage: updateItem,
-    qty:saleslabouramount,labouramount,    
-    custom_labour_amount: tatoalamount,
-    custom_fine_weight: finevalue,
-    custom_sales_labour_type: saleslabouramount,
-    custom_labour_type: labouramount,
-    custom_total_amount: calculateTotal,   
-    
+frappe.ui.form.on('Sales Order', {
+    custom_wholesale: function(frm) {
+        frm.doc.items.forEach(function(row) {
+            var custom_gold_rate = row.custom_gold_rate;
+            var custom_westage = row.custom_westage;  
+            var custom_net_weight = row.custom_net_weight   
+            var gold_value = 0;
+            if(frm.doc.custom_wholesale == 1){
+                gold_value = custom_gold_rate * custom_net_weight + custom_westage;
+            }
+            else{
+                gold_value = custom_gold_rate * custom_net_weight;
+            }
+            frappe.model.set_value(row.doctype, row.name, 'custom_gold_value', gold_value); 
+       });
+       
+    }
 
-   
+})
+frappe.ui.form.on('Sales Order Item', {
+    
+        item_code: function(frm, cdt, cdn){
+            fetchGoldRate(frm, cdt, cdn);
+        },
+        custom_gross_weight: function(frm, cdt, cdn) {
+            updateItem(frm, cdt, cdn);
+        },
+        custom_less_weight: function(frm, cdt, cdn) {
+            updateItem(frm, cdt, cdn);
+        }, 
+        custom_westage: function(frm, cdt, cdn) {
+            updateItem(frm, cdt, cdn);
+        },                  
+        custom_purity_percentage: function(frm, cdt, cdn) {
+            updateItem(frm, cdt, cdn);
+        }, 
+        qty: function(frm, cdt, cdn) {
+            saleslabouramount(frm, cdt, cdn);
+            labouramount(frm, cdt, cdn);
+        },  
+        custom_sales_labour_amount: function(frm, cdt, cdn){
+            tatoalamount(frm, cdt, cdn);
+        },
+        custom_gold_value: function(frm, cdt, cdn){
+            tatoalamount(frm, cdt, cdn);
+        },
+        custom_other_amount: function(frm, cdt, cdn){
+            tatoalamount(frm, cdt, cdn);
+        },
+         
+        custom_sales_labour_type: function(frm, cdt, cdn){
+            saleslabouramount(frm, cdt, cdn);
+        },
 
 });
+
+
 //-----calculate custom net weight , custom fine weight , custom gold value -------//
 function updateItem(frm, cdt, cdn) {
     var child_doc = locals[cdt][cdn];
@@ -24,12 +63,18 @@ function updateItem(frm, cdt, cdn) {
     var custom_less_weight = child_doc.custom_less_weight;
     var custom_purity_percentage = child_doc.custom_purity_percentage;
     var custom_gold_rate = child_doc.custom_gold_rate;
+    var custom_westage = child_doc.custom_westage;
     
 
     var custom_net_weight = custom_gross_weight - custom_less_weight;
     var fine_weight = custom_net_weight / (custom_purity_percentage / 100);
-    var gold_value = custom_gold_rate * custom_net_weight;
-    
+    var gold_value = 0;
+    if(frm.doc.custom_wholesale == 1){
+        gold_value = custom_gold_rate * custom_net_weight + custom_westage;
+    }
+    else{
+        gold_value = custom_gold_rate * custom_net_weight;
+    }   
     
 
     frappe.model.set_value(cdt, cdn, 'custom_net_weight', custom_net_weight);
@@ -45,20 +90,20 @@ function updateItem(frm, cdt, cdn) {
 
 
 //-----calculate custom net weight , custom fine weight-------//
-function finevalue(frm,cdt,cdn){
-    var child_doc = locals[cdt][cdn];
+// function finevalue(frm,cdt,cdn){
+//     var child_doc = locals[cdt][cdn];
 
-    var custom_gold_rate = child_doc.custom_gold_rate;
-    var custom_fine_weight = child_doc.custom_fine_weight;
+//     var custom_gold_rate = child_doc.custom_gold_rate;
+//     var custom_fine_weight = child_doc.custom_fine_weight;
 
-    var fine_value = custom_fine_weight * custom_gold_rate;  
+//     var fine_value = custom_fine_weight * custom_gold_rate;  
 
-    if(custom_fine_weight !== undefined && custom_gold_rate !== undefined){
+//     if(custom_fine_weight && custom_gold_ratev){
               
-        frappe.model.set_value(cdt, cdn, 'custom_fine_value', fine_value);
+//         frappe.model.set_value(cdt, cdn, 'custom_fine_value', fine_value);
         
-    }
-}
+//     }
+// }
 
 
 //-----calculate custom amount------//
@@ -67,13 +112,13 @@ function tatoalamount(frm, cdt, cdn){
 
     var custom_gold_value = child_doc.custom_gold_value;
     var custom_other_amount = child_doc.custom_other_amount;
-    var custom_labour_amount = child_doc.custom_labour_amount;
+    var custom_sales_labour_amount = child_doc.custom_sales_labour_amount;
 
-    var total_amount = custom_gold_value + custom_other_amount + custom_labour_amount;
-    if(custom_gold_value && custom_other_amount && custom_labour_amount){
-        frappe.model.set_value(cdt, cdn, 'custom_total_amount', total_amount);
+    var total_amount = custom_gold_value + custom_other_amount + custom_sales_labour_amount;
+    
+    frappe.model.set_value(cdt, cdn, 'custom_total_amount', total_amount);
         
-    }
+
 
 }
 
@@ -81,69 +126,40 @@ function tatoalamount(frm, cdt, cdn){
 function saleslabouramount(frm, cdt, cdn){
     var child_doc = locals[cdt][cdn];
 
-    var custom_value_added = child_doc.custom_value_added;
+    var custom_sales_labour_rate = child_doc.custom_sales_labour_rate;
     var quantity = child_doc.qty;
     var salestype = child_doc.custom_sales_labour_type;
-    var custom_fine_value = child_doc.custom_fine_value;
+    var custom_gold_rate = child_doc.custom_gold_rate;
+    var custom_fine_weight = child_doc.custom_fine_weight;
     var custom_net_weight = child_doc.custom_net_weight;
     var custom_gross_weight = child_doc.custom_gross_weight;
-   
-     
+
+    var fine_value = custom_fine_weight * custom_gold_rate;
+    
 
     if(salestype == 'Flat'){
-        var flat = custom_value_added * quantity;
+        var flat = custom_sales_labour_rate * quantity;
         frappe.model.set_value(cdt, cdn, 'custom_sales_labour_amount', flat);
+       
     }
 
     if(salestype == 'On Gold Value Percentage'){
-        var Percentage = custom_fine_value * (custom_value_added / 100);
+        var Percentage = fine_value * (custom_sales_labour_rate / 100);
         frappe.model.set_value(cdt, cdn, 'custom_sales_labour_amount', Percentage);
     }
 
     if(salestype == 'On Gross Weight Per Gram'){
-        var Percentage = custom_gross_weight * custom_value_added;
+        var Percentage = custom_gross_weight * custom_sales_labour_rate;
         frappe.model.set_value(cdt, cdn, 'custom_sales_labour_amount', Percentage);
     }
 
     if(salestype == 'On Net Weight Per Gram'){
-        var netweight =  custom_net_weight * custom_value_added;
+        var netweight =  custom_net_weight * custom_sales_labour_rate;
         frappe.model.set_value(cdt, cdn, 'custom_sales_labour_amount', netweight);
     }   
 }
 
 
-//-----calculate custom labour amountt------//
-function labouramount(frm, cdt, cdn){
-    var child_doc = locals[cdt][cdn];
-
-    var custom_labour_rate = child_doc.custom_labour_rate;
-    var quantity = child_doc.qty;
-    var salestype = child_doc.custom_labour_type;
-    var custom_fine_value = child_doc.custom_fine_value;
-    var custom_net_weight = child_doc.custom_net_weight;
-    var custom_gross_weight = child_doc.custom_gross_weight;   
-       
-
-    if(salestype == 'Flat'){
-        var flat =  custom_labour_rate * quantity;
-        frappe.model.set_value(cdt, cdn, 'custom_labour_amount', flat);
-    }
-
-    if(salestype == 'On Gold Value Percentage'){
-        var Percentage = custom_fine_value * (custom_labour_rate / 100);
-        frappe.model.set_value(cdt, cdn, 'custom_labour_amount', Percentage);
-    }
-
-    if(salestype == 'On Gross Weight Per Gram'){
-        var Percentage = custom_gross_weight * custom_labour_rate;
-        frappe.model.set_value(cdt, cdn, 'custom_labour_amount', Percentage);
-    }
-
-    if(salestype == 'On Net Weight Per Gram'){
-        var netweight =  custom_net_weight * custom_labour_rate;
-        frappe.model.set_value(cdt, cdn, 'custom_labour_amount', netweight);
-    }
-}
 
 
 //-----fetch Gold Rate------//
@@ -169,20 +185,4 @@ function fetchGoldRate(frm, cdt, cdn) {
         }
     });
 }
-//-----set all Total in purchase order------//
-function  calculateTotal(frm){
-    var net_weight = 0, fine_weight = 0, gross_weight = 0,less_weight = 0;
-    frm.doc.items.forEach(function(row) {
-        net_weight += row.custom_net_weight;
-        fine_weight += row.custom_fine_weight;
-        gross_weight += row.custom_gross_weight;
-        less_weight += row.custom_less_weight;
-        
-                  
-    });       
-    frm.set_value('custom_total_net_weight', net_weight);
-    frm.set_value('custom_total_fine_weight', fine_weight);
-    frm.set_value('custom_total_gross_weight', gross_weight);
-    frm.set_value('custom_total_less_weight', less_weight);
-     
-}
+
